Cleaned up controller test classes
